Added exit logic when pressing "q"

+++ PersonalFinancialTracker/src/main/java/dev/personal/financial/tracker/UI/handler/AdminHandler.java
@@ -16,12 +16,20 @@ public class AdminHandler {
 
     public void blockUser() {
         String userId = printer.readNonEmptyString("Введите id пользователя, которого хотите заблокировать:");
+        if (userId == null) {
+            printer.printInfo("Блокировка пользователя отменена.");
+            return;
+        }
         adminController.blockUser(userId);
         printer.printSuccess("Пользователь заблокирован");
     }
 
     public void deleteUser() {
         String userId = printer.readNonEmptyString("Введите id пользователя, которого хотите удалить:");
+        if (userId == null) {
+            printer.printInfo("Удаление пользователя отменено.");
+            return;
+        }
         adminController.deleteUser(userId);
         printer.printSuccess("Пользователь удален");
     }
